+++ staramr/blast/results/plasmidfinder/PlasmidfinderHitHSP.py
@@ -1,6 +1,7 @@
 import logging
 import re
 import pandas as pd
+from typing import List
 
 from staramr.blast.results.AMRHitHSP import AMRHitHSP
 
@@ -25,12 +26,17 @@ class PlasmidfinderHitHSP(AMRHitHSP):
         logger.debug("record=%s", self._blast_record)
 
         splitList = re.split('_', self.get_amr_gene_id())
-        re_search = list(filter(None, splitList))
+        re_search = list(filter(None, splitList)) # type: List[str]
 
         if not re_search:
             raise Exception("Could not split up seq name for [" + self.get_amr_gene_id() + "]")
 
         length = len(re_search)
+
+        # Add empty string if gene variant is missing
+        if length == 2:
+            re_search.insert(1,'')
+            length += 1
 
         if 3 <= length <= 5:
             self._gene = re_search[0]