+++ staramr/results/AMRDetectionSummary.py
@@ -15,6 +15,7 @@ Summarizes both ResFinder, PointFinder, and PlasmidFinder database results into 
 
 class AMRDetectionSummary:
     SEPARATOR = ','
+    FLOAT_DECIMALS = 2
 
     def __init__(self, files, resfinder_dataframe: DataFrame, pointfinder_dataframe=None, plasmidfinder_dataframe=None) -> None:
         """
@@ -68,9 +69,8 @@ class AMRDetectionSummary:
         negative_entries = None
 
         if len(negative_res_names_set) != len(names_set) or resistance_frame.empty:
-            logger.debug("Went here")
-            negative_resistance_entries = pd.DataFrame([[x, 'None', 'Sensitive'] for x in negative_res_names_set],
-                                                       columns=('Isolate ID', 'Gene', 'Predicted Phenotype')).set_index('Isolate ID')
+            negative_resistance_entries = pd.DataFrame([[x, 'None', 'Sensitive', '', ''] for x in negative_res_names_set],
+                                                       columns=('Isolate ID', 'Gene', 'Predicted Phenotype', 'Start', 'End')).set_index('Isolate ID')
             negative_resistance_entries['Data Type']='Resistance'
             negative_entries = negative_resistance_entries
 
@@ -115,7 +115,13 @@ class AMRDetectionSummary:
         if plasmid_frame is not None:
             plasmid_frame = self._compile_plasmids(plasmid_frame)
 
-            resistance_frame = resistance_frame.merge(plasmid_frame, on='Isolate ID', how='left').fillna(value={'Plasmid Genes': 'None'})
+            if resistance_frame.empty:
+                resistance_frame = resistance_frame.append(plasmid_frame)
+                resistance_frame = resistance_frame.fillna(value={'Genotype': 'None', 'Predicted Phenotype': 'Sensitive'})
+            else:
+                resistance_frame = resistance_frame.merge(plasmid_frame, on='Isolate ID', how='left').fillna(value={'Plasmid Genes': 'None'})
+
+            resistance_frame = resistance_frame.reindex(columns=['Genotype', 'Predicted Phenotype', 'Plasmid Genes'])
 
         return resistance_frame.sort_index()
 
@@ -125,6 +131,7 @@ class AMRDetectionSummary:
         else:
             resistance_frame = self._resfinder_dataframe.copy()
             resistance_frame['Data Type']='Resistance'
+            resistance_frame = resistance_frame.round({'%Identity': self.FLOAT_DECIMALS, '%Overlap': self.FLOAT_DECIMALS})
 
         if self._plasmidfinder_dataframe is None:
             plasmid_frame = None
@@ -139,6 +146,7 @@ class AMRDetectionSummary:
             else:
                 point_frame = self._pointfinder_dataframe.copy()
                 point_frame['Data Type']='Resistance'
+                point_frame = point_frame.round({'%Identity': self.FLOAT_DECIMALS, '%Overlap': self.FLOAT_DECIMALS})
                 point_frame = point_frame.reindex(columns=column_names)
 
             if resistance_frame is not None:
@@ -155,6 +163,7 @@ class AMRDetectionSummary:
         if plasmid_frame is not None:
             plasmid_frame['Data Type']='Plasmid'
             plasmid_frame['Predicted Phenotype']=''
+            plasmid_frame = plasmid_frame.round({'%Identity': self.FLOAT_DECIMALS, '%Overlap': self.FLOAT_DECIMALS})
 
             if resistance_frame is not None:
                 resistance_frame = resistance_frame.append(plasmid_frame, sort=True)